Move stack up to be shared with core
